requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beanie==1.17.0
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
lazy-model==0.0.5
motor==3.1.1
multidict==6.0.4
numpy==1.24.2
py-cord==2.4.1
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
toml==0.10.2
typing_extensions==4.5.0
yarl==1.8.2
